==> notebooks/03-lv-features-selection.ipynb <==
%load_ext autoreload
%autoreload 2

import os
import sys
module_path = os.path.abspath(os.path.join('..')) # or the path to your source code
sys.path.insert(0, module_path)

import pandas as pd
from tqdm import tqdm
import re
import nltk
import seaborn as sns
import matplotlib.pyplot as plt

from nlp_project.config import PROCESSED_DATA_DIR, INTERIM_DATA_DIR
from nlp_project.plots import *
from nlp_project.dataset import build_ontolgy

dataset = pd.read_csv(PROCESSED_DATA_DIR / 'preprocessed_dataset.csv', sep=';')
dataset

descr = dataset[['description', 'asin', 'brand']]
descr.drop_duplicates(inplace=True)
descr

brands_product_count = descr.groupby('brand').size().reset_index(name='counts').sort_values(by='counts', ascending=False)
brands_product_count.to_csv(INTERIM_DATA_DIR / 'brand_product_count.csv',sep=';', index=False)
brands_product_count

import plotly.graph_objects as go

labels = brands_product_count['brand']
values = brands_product_count['counts']

pie_chart_go(labels=labels, values=values)

import nltk
nltk.download('punkt')
nltk.download('wordnet')
nltk.download('stopwords')
nltk.download('averaged_perceptron_tagger')
nltk.download('tagsets')
nltk.download('universal_tagset')

from nltk.corpus import stopwords
stop_words = stopwords.words('english')

from nltk.stem import WordNetLemmatizer
lemmatizer = WordNetLemmatizer()

'''
def build_ontolgy(df):
    nouns = []
    for index, row in tqdm(df.iterrows(), total = len(df)):
        text = row['description']
        text_cleaned = re.sub(r'[^\w\s]', '', text) #cleaning
        words = nltk.word_tokenize(text_cleaned) #tokenization
        words = [w.lower() for w in words] #lower casing
        tagged_words = nltk.pos_tag(words, tagset='universal') #POS-tagging
        preprocessed_words = [(lemmatizer.lemmatize(w), tag) for w, tag in tagged_words if not w in stop_words] #stopwords removal and lemmatization
        #print(filter_text)
        for word, tag in preprocessed_words:
            if(tag == 'NOUN'):
                nouns.append(word)
            
    return nouns
'''

ont = build_ontolgy(descr, lemmatizer, stop_words)

from collections import Counter

nouns_freq = Counter(ont)
#print(value_counts)
# Convert Counter to DataFrame
df_nouns_freq = pd.DataFrame(nouns_freq.items(), columns=['value', 'count'])
df_nouns_freq = df_nouns_freq.sort_values(by='count', ascending=False)
df_nouns_freq.head(30)

sb_bar_plot(x = df_nouns_freq['count'][:25],
            y = df_nouns_freq['value'][:25],
            orient='h',
            title='Most frequent nouns in description', 
            xlabel='Count', 
            ylabel='Noun')

Esporto e analizzo in excel manualmente

df_nouns_freq.to_csv(PROCESSED_DATA_DIR / 'ontology.csv', index=False, sep=';')

# Parole chiave da cercare
#features = list(pd.read_csv('data/ontology_filtered.csv', sep=';')['value'])


#features

# Crea un'espressione regolare che combini tutte le parole chiave
#pattern = '|'.join(features)

# Filtra il DataFrame
#filtered_df = dataset[dataset['preprocessedText'].str.contains(pattern, case=False, na=False)]

#filtered_df.loc[0,'preprocessedText']

#filtered_df.to_csv('data/preprocessed_dataset.csv', index=False, sep=';')
